miwear: comprehensive code quality refactoring
MIWEARRTOS-113

Refactor the entire codebase for better maintainability, type safety,
and cross-platform compatibility:

- Extract shared APP_NAME_TO_ID/APP_ID_TO_NAME into constants.py,
  remove duplicate html.py (identical copy of loganalyzer.py)
- Remove legacy setup.py and unused ymodem.py module
- Replace all shell commands in log.py with Python stdlib (shutil,
  tarfile, gzip, os.chmod), remove ShellRunner class entirely
- Replace subprocess tar extraction in targz.py with tarfile module
- Fix assert.py StopIteration crash with friendly error handling
- Split monolithic check.py (1800+ lines) into check/ package with
  scanner, tui, report, and __main__ modules
- Add type annotations across all modules
- Unify --version to use argparse action='version' pattern
- Fix Highlight.Convert to check isatty() before ANSI output
- Rename CHECK_ERROR_EXIT to check_error_exit (PEP 8)
- Consolidate test/ into tests/data/, add pytest suite (21 tests)
- Add mypy and pytest to CI workflow
